vite запускает проект
