+++ src/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Routes, Route } from "react-router-dom";
-import { Home } from './pages/Home.jsx';
+import { Home } from './pages/Home';
 import "./styles.css"
 
 export default function App() {
