+++ src/components/RedButton.tsx
@@ -1,0 +1,5 @@
+export const RedButton = () => {
+    return(
+        <button>go</button>
+    )
+}
