+++ src/icons/Cart.tsx
@@ -1,0 +1,11 @@
+export const Cart = () => {
+    return (
+        <svg width="20.000000" height="20.000000" viewBox="0 0 24 24" fill="none">
+
+	<path id="Vector" d="M20 7.14L16.6 7.14L13.03 0.34C12.85 0 12.47 -0.11 12.17 0.1C11.88 0.3 11.78 0.74 11.96 1.08L15.14 7.14L4.85 7.14L8.03 1.08C8.21 0.74 8.11 0.3 7.82 0.1C7.52 -0.11 7.14 0 6.96 0.34L3.39 7.14L0 7.14L0 8.57L1.35 8.57L2.94 18.25C3.11 19.26 3.88 20 4.78 20L15.21 20C16.11 20 16.88 19.26 17.05 18.25L18.64 8.57L20 8.57C20 8.57 20 7.14 20 7.14ZM15.82 17.98C15.77 18.32 15.51 18.57 15.21 18.57L4.78 18.57C4.48 18.57 4.22 18.32 4.17 17.98L2.62 8.57L17.37 8.57L15.82 17.98Z" fill="#FFFFFF" fillOpacity="1.000000" fillRule="nonzero"/>
+	<rect id="counter" rx="5.500000" width="11.000000" height="11.000000" transform="translate(12.500000 12.500000)" fill="#F14F4F" fillOpacity="1.000000"/>
+	<path id="1" d="M16.83 15.58L16.83 16.28L18.07 15.91L18.07 20.78L18.75 20.78L18.75 15.1L16.83 15.58Z" fill="#FFFFFF" fillOpacity="1.000000" fillRule="evenodd"/>
+</svg>
+
+    )
+}
