+++ src/icons/LogoImg.tsx
@@ -1,0 +1,8 @@
+export const LogoImg = () => {
+    return (
+        <svg width="160.299072" height="30.000000" viewBox="0 0 160.299 30" fill="none" >
+            <defs />
+            <path id="Logo" d="M75.57 -2.68e-5L71.04 -2.68e-5L71.04 11.81C70.59 11.13 70.04 10.58 69.39 10.17C68.51 9.58 67.49 9.29 66.33 9.29C65.21 9.29 64.15 9.57 63.15 10.14C62.15 10.7 61.33 11.51 60.69 12.57C60.07 13.63 59.76 14.94 59.76 16.5C59.76 18.04 60.07 19.35 60.69 20.43Q61.45 21.68 62.54 22.47Q62.83 22.68 63.15 22.86C64.15 23.42 65.21 23.7 66.33 23.7C67.49 23.7 68.51 23.41 69.39 22.83C70.04 22.41 70.59 21.86 71.04 21.18L71.04 23.4L75.57 23.4L75.57 -2.68e-5ZM71.04 16.5C71.04 15.76 70.88 15.13 70.56 14.61C70.24 14.07 69.83 13.66 69.33 13.38Q68.8 13.06 68.23 12.98Q68.03 12.96 67.83 12.96C67.23 12.96 66.68 13.11 66.18 13.41C65.68 13.69 65.28 14.1 64.98 14.64C64.68 15.16 64.53 15.78 64.53 16.5C64.53 17.21 64.68 17.85 64.98 18.39C65.28 18.91 65.68 19.32 66.18 19.62C66.68 19.9 67.23 20.04 67.83 20.04C68.35 20.04 68.85 19.9 69.33 19.62C69.83 19.34 70.24 18.94 70.56 18.42C70.88 17.87 71.04 17.24 71.04 16.5ZM11.61 11.91L11.61 15.81L17.19 15.81C17.01 16.35 16.76 16.86 16.44 17.34C16.14 17.79 15.77 18.2 15.33 18.57C14.89 18.91 14.36 19.18 13.74 19.37Q12.81 19.65 11.67 19.65Q9.92 19.65 8.54 18.93Q8.45 18.88 8.36 18.84C7.42 18.28 6.67 17.5 6.11 16.5Q5.3 14.97 5.3 12.9Q5.3 10.8 6.14 9.35Q6.71 8.4 7.51 7.74Q7.95 7.38 8.45 7.1C9.43 6.58 10.51 6.32 11.67 6.32C12.89 6.32 13.9 6.6 14.7 7.16C15.52 7.7 16.17 8.39 16.65 9.23L21.15 6.89C20.49 5.91 19.7 5.05 18.78 4.31Q18.73 4.28 18.69 4.24Q17.37 3.15 15.66 2.54C14.48 2.12 13.15 1.91 11.67 1.91Q9.17 1.91 7.04 2.69Q5.18 3.38 3.74 4.56Q3.53 4.73 3.32 4.91C2.26 5.87 1.44 7.03 0.86 8.39Q-1.95e-5 10.41 -1.95e-5 12.9Q-1.96e-5 15.39 0.83 17.45Q1.47 18.93 2.44 20.13Q2.81 20.58 3.23 21Q3.43 21.18 3.64 21.36Q5.04 22.57 6.86 23.28C8.26 23.8 9.8 24.06 11.49 24.06C13.27 24.06 14.85 23.75 16.23 23.13C17.61 22.49 18.76 21.61 19.68 20.49C20.6 19.37 21.28 18.08 21.72 16.62C22.16 15.16 22.33 13.59 22.23 11.91L11.61 11.91ZM103.6 2.39L91.42 19.68L102.1 19.68L102.1 23.4L107.38 23.4L107.38 19.68L110.68 19.68L110.68 15.36L107.38 15.36L107.38 2.39L103.6 2.39ZM102.67 16.02L102.67 9.74L98.14 16.02L102.67 16.02ZM25.22 12.97Q24.37 14.49 24.37 16.5Q24.37 18.39 25.15 19.86Q25.26 20.07 25.39 20.28Q25.72 20.81 26.14 21.26Q27 22.19 28.21 22.8C29.41 23.4 30.76 23.7 32.26 23.7C33.76 23.7 35.1 23.4 36.28 22.8C37.48 22.2 38.42 21.36 39.1 20.28C39.8 19.2 40.15 17.94 40.15 16.5C40.15 15.04 39.8 13.77 39.1 12.69C38.42 11.61 37.48 10.77 36.28 10.17C35.1 9.56 33.76 9.26 32.26 9.26C30.76 9.26 29.41 9.56 28.21 10.17C27.01 10.77 26.07 11.61 25.39 12.69Q25.3 12.82 25.22 12.97ZM42.92 12.97Q42.06 14.49 42.06 16.5Q42.06 18.39 42.85 19.86Q42.96 20.07 43.08 20.28C43.76 21.36 44.7 22.2 45.9 22.8C47.1 23.4 48.45 23.7 49.95 23.7C51.45 23.7 52.79 23.4 53.97 22.8C55.17 22.2 56.11 21.36 56.79 20.28C57.49 19.2 57.84 17.94 57.84 16.5C57.84 15.04 57.49 13.77 56.79 12.69C56.11 11.61 55.17 10.77 53.97 10.17C52.79 9.56 51.45 9.26 49.95 9.26C48.45 9.26 47.1 9.56 45.9 10.17C44.7 10.77 43.76 11.61 43.08 12.69Q43 12.82 42.92 12.97ZM128.96 12.97Q128.11 14.49 128.11 16.5Q128.11 18.39 128.89 19.86Q129 20.07 129.13 20.28Q129.46 20.81 129.89 21.26Q130.74 22.19 131.95 22.8C133.15 23.4 134.5 23.7 136 23.7Q138.25 23.7 140.02 22.8C141.22 22.2 142.16 21.36 142.84 20.28C143.54 19.2 143.89 17.94 143.89 16.5C143.89 15.04 143.54 13.77 142.84 12.69C142.16 11.61 141.22 10.77 140.02 10.17C138.84 9.56 137.5 9.26 136 9.26C134.5 9.26 133.15 9.56 131.95 10.17C130.75 10.77 129.81 11.61 129.13 12.69Q129.04 12.82 128.96 12.97ZM81.72 19.77C81.16 19.43 80.53 18.96 79.83 18.39L77.94 21.39C78.56 21.87 79.23 22.31 79.95 22.71C80.67 23.09 81.45 23.39 82.29 23.61C83.15 23.85 84.08 23.97 85.08 23.97C86.72 23.97 87.99 23.6 88.89 22.86C89.81 22.1 90.27 20.98 90.27 19.5C90.27 18.64 90.08 17.95 89.7 17.43C89.34 16.91 88.85 16.48 88.23 16.14C87.63 15.8 86.97 15.5 86.25 15.24C85.57 14.98 84.99 14.77 84.51 14.61C84.05 14.43 83.69 14.24 83.43 14.04C83.19 13.84 83.07 13.59 83.07 13.29C83.07 13.03 83.17 12.82 83.37 12.66C83.59 12.5 83.91 12.42 84.33 12.42C84.95 12.42 85.63 12.6 86.37 12.96C87.11 13.3 87.77 13.74 88.35 14.28L89.97 11.34C89.43 10.94 88.85 10.59 88.23 10.29C87.63 9.98 86.98 9.74 86.28 9.56C85.6 9.38 84.86 9.29 84.06 9.29Q82.56 9.29 81.39 9.74Q80.42 10.13 79.77 10.82Q79.66 10.94 79.56 11.07C79.12 11.65 78.9 12.36 78.9 13.2C78.9 14.08 79.09 14.8 79.47 15.36C79.85 15.92 80.33 16.37 80.91 16.7C81.51 17.03 82.14 17.31 82.8 17.54C83.54 17.77 84.12 17.98 84.54 18.18C84.98 18.37 85.29 18.59 85.47 18.81C85.65 19.01 85.74 19.26 85.74 19.56Q85.74 19.85 85.62 20.07Q85.54 20.22 85.41 20.34C85.19 20.54 84.85 20.64 84.39 20.64C84.07 20.64 83.69 20.58 83.25 20.45C82.81 20.32 82.3 20.09 81.72 19.77ZM118.84 30L127.84 9.59L123.19 9.59L119.94 17.74L116.65 9.59L111.67 9.59L117.64 22.29L114.28 30L118.84 30ZM151.8 19.89C151.5 19.43 151.35 18.79 151.35 18L151.35 9.59L146.61 9.59L146.61 18.48Q146.61 20.94 147.78 22.35C148.58 23.27 149.84 23.73 151.56 23.73C152.58 23.73 153.43 23.52 154.11 23.1C154.79 22.66 155.34 22.07 155.76 21.33L155.76 23.4L160.29 23.4L160.29 9.59L155.76 9.59L155.76 18C155.76 18.54 155.66 19 155.46 19.37C155.26 19.76 154.98 20.04 154.62 20.25C154.28 20.45 153.86 20.55 153.36 20.55C152.62 20.55 152.1 20.33 151.8 19.89ZM29.47 18.36C29.17 17.84 29.02 17.21 29.02 16.5C29.02 15.78 29.17 15.16 29.47 14.64Q29.82 14.02 30.34 13.63Q30.49 13.52 30.64 13.44C31.12 13.16 31.66 13.02 32.26 13.02C32.84 13.02 33.37 13.16 33.85 13.44Q34.15 13.6 34.4 13.83Q34.77 14.17 35.02 14.64C35.32 15.16 35.47 15.78 35.47 16.5C35.47 17.21 35.32 17.84 35.02 18.36C34.74 18.86 34.35 19.25 33.85 19.53C33.37 19.81 32.84 19.95 32.26 19.95C31.66 19.95 31.12 19.81 30.64 19.53C30.16 19.25 29.77 18.86 29.47 18.36ZM47.16 18.36C46.86 17.84 46.71 17.21 46.71 16.5C46.71 15.78 46.86 15.16 47.16 14.64Q47.52 14.02 48.04 13.63Q48.18 13.52 48.33 13.44C48.81 13.16 49.35 13.02 49.95 13.02C50.53 13.02 51.06 13.16 51.54 13.44Q51.85 13.6 52.1 13.83Q52.46 14.17 52.71 14.64C53.01 15.16 53.16 15.78 53.16 16.5C53.16 17.21 53.01 17.84 52.71 18.36C52.43 18.86 52.04 19.25 51.54 19.53C51.06 19.81 50.53 19.95 49.95 19.95C49.35 19.95 48.81 19.81 48.33 19.53C47.85 19.25 47.46 18.86 47.16 18.36ZM133.21 18.36C132.91 17.84 132.76 17.21 132.76 16.5C132.76 15.78 132.91 15.16 133.21 14.64Q133.56 14.02 134.08 13.63Q134.23 13.52 134.38 13.44C134.86 13.16 135.4 13.02 136 13.02C136.58 13.02 137.11 13.16 137.59 13.44Q137.89 13.6 138.14 13.83Q138.51 14.17 138.76 14.64C139.06 15.16 139.21 15.78 139.21 16.5C139.21 17.21 139.06 17.84 138.76 18.36C138.48 18.86 138.09 19.25 137.59 19.53C137.11 19.81 136.58 19.95 136 19.95C135.4 19.95 134.86 19.81 134.38 19.53C133.9 19.25 133.51 18.86 133.21 18.36Z" fill="#FFFFFF" fillOpacity="1.000000" fillRule="evenodd" />
+        </svg>
+    )
+}
