+++ src/index.jsx
@@ -1,13 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-
 import { BrowserRouter } from "react-router-dom";
-import App from './App.js';
-
+import App from './App.jsx';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <BrowserRouter>
     <App />
-  </BrowserRouter>)
+  </BrowserRouter>
+  )
 ;
