--- src/pages/Home.jsx
@@ -1,12 +1,0 @@
-import { Headers } from "../components/header"
-import { Footer } from "../components/footer"
-
-export const Home = () =>{
-    return(
-        <>
-        <Headers />
-        <Footer />
-        </>
-    
-    )
-}
