+++ vite.config.ts
@@ -4,7 +4,14 @@ import react from '@vitejs/plugin-react'
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
+  server: {
+    port: 3000,
+    hmr: true
+  },
   resolve: {
+    alias: {
+      src: "/src"
+    } ,
     extensions: ['.js','.jsx','.ts', '.tsx', '.json', '.css']
   }
 })
